Move kit add modal into a separate page
