+++ inventory/kits/templates/kit_create.html
@@ -1,0 +1,6 @@
+{% extends "inventory_base.html" %} {% load static %} {% block title %}Kit
+Index{% endblock %} {% block content %}
+
+<div id="root" style="width: 100vw; height: 100vh"></div>
+<script src="{% static 'webpack/kitAdd.js' %}"></script>
+{%endblock%}
