+++ inventory/kits/views.py
@@ -42,6 +42,11 @@ def cart(request):
     return render(request, "kit_cart.html")
 
 
+@login_required(login_url="/r'^login/$'")
+def kit_create(request):
+    return render(request, "kit_create.html")
+
+
 @api_view(["GET"])
 @permission_classes([IsAuthenticated])
 def api_kits(request):
