+++ inventory/src/components/KitAdd/schema.js
