+++ inventory/src/pages/index.js
@@ -8,3 +8,4 @@ export * from './itemList.js';
 export * from './kitIndex.js';
 export * from './createBlueprint.js';
 export * from './kitCart.js';
+export * from './kitAdd.js';
