+++ inventory/src/pages/render/kitAdd.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import ReactDOM from 'react-dom/client';
+
+import { InventoryProvider } from '../../providers';
+import { KitAdd } from '../kitAdd';
+
+const root = ReactDOM.createRoot(document.querySelector('#root'));
+root.render(
+  <InventoryProvider>
+    <KitAdd />
+  </InventoryProvider>,
+);
